+++ xenon-common/src/main/java/com/vmware/xenon/common/ReflectionUtils.java
@@ -47,4 +47,16 @@ public class ReflectionUtils {
             throw new RuntimeException(e);
         }
     }
+
+    /**
+     * Checks if fieldName is present and accessible in service type
+     */
+    public static boolean hasField(Class<? extends Service> type, String fieldName) {
+        try {
+            type.getField(fieldName);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
 }