+++ xenon-common/src/main/java/com/vmware/xenon/common/Utils.java
@@ -941,4 +941,5 @@ public class Utils {
 
         return result;
     }
+
 }